+++ audiopoli_dashboard/pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   google_maps_flutter_platform_interface: ^2.4.3
   google_maps_flutter_web: ^0.5.4+3
   sticky_headers: ^0.3.0
+  intl: ^0.19.0
 
 dev_dependencies:
   flutter_test:
